rename noop -> noopLegacy
