+++ adapter/noopLegacy/BUILD
