+++ adapter/noopLegacy/noop.go
@@ -14,7 +14,7 @@
 
 // Package noop is an empty adapter implementing every aspect.
 // WARNING: Not intended for actual use. This is a stand-in adapter used in benchmarking Mixer's adapter framework.
-package noop
+package noopLegacy
 
 import (
 	rpc "github.com/googleapis/googleapis/google/rpc"
